Trait hotg_rune_codegen::Environment

pub trait Environment {
    fn compile(&mut self, project: Project) -> Result<Vec<u8>, Error>;

    fn read_file(&mut self, filename: &Path) -> Result<Vec<u8>, Error>;

    fn build_info(&self) -> Option<Value> { ... }
}

Required methods

fn compile(&mut self, project: Project) -> Result<Vec<u8>, Error>

Compile a Rust project to WebAssembly, returning the contents of the compiled binary.

fn read_file(&mut self, filename: &Path) -> Result<Vec<u8>, Error>

Read a file from the file system, relative to the “current” directory.

Provided methods

fn build_info(&self) -> Option<Value>

Get a JSON object which contains information about the program compiling this Rune.